removed TrackResults Component and worked on styling
