+++ src/components/App/App.css
@@ -15,7 +15,7 @@ h1 {
   background-color: #b16d22;
   text-align: center;
   font-family: 'Montserrat', sans-serif;
-  font-size: 1.88rem;
+  font-size: 1.5rem;
   color: #fff;
 }
 
@@ -30,26 +30,26 @@ a {
 
 h2 {
   font-family: "Montserrat", sans-serif;
-  font-size: 1.55rem;
+  font-size: 1rem;
 }
 
 .App {
   height: 100%;
-  padding: 0 17% 10% 17%;
-  background-image: url("https://source.unsplash.com/8e0EHPUx3Mo");
+  padding: 0 10% 5% 10%;
+  /* background-image: url("https://source.unsplash.com/8e0EHPUx3Mo");
   background-size: cover;
   background-position: center;
-  background-repeat: no-repeat;
+  background-repeat: no-repeat; */
   font-family: "Montserrat", sans-serif;
   font-weight: 500;
   color: #fff;
 }
 
-.App-playlist {
+/* .App-playlist {
   display: flex;
   justify-content: space-between;
   width: 100%;
-}
+} */
 
 @media only screen and (max-width: 1020px) {
   .App-playlist {
