+++ src/components/Home/Home.css
@@ -5,7 +5,7 @@
 }
 
 .inner_container {
-  margin-left: 1rem;
+  margin-left: 2rem;
 }
 
 h4 {
