+++ src/components/SearchResults/SearchResults.css
@@ -1,13 +1,14 @@
 .SearchResults {
-  /* width: 50%; */
+  width: 50%;
   /* height: 950px; */
   /* overflow-y: scroll; */
-  padding: 0.88rem;
-  background-color: rgba(228, 194, 194, 0.9);
-  box-shadow: 0 4px 2px 2px #000000;
+  padding: 1rem;
+  background-color: rgba(228, 194, 194, 0.5);
+  box-shadow: 0 0.1rem .2rem black;
   color: black;
-  margin-right: 1rem;
+  margin: 0 auto;
 }
+
 ::-webkit-scrollbar {
   width: 0px;
   background: transparent;
