+++ src/components/Track/Track.css
@@ -1,14 +1,14 @@
 .Track {
   display: flex;
   align-items: center;
-  border-bottom: 1px solid rgba(256, 256, 256, 0.8);
+  border-bottom: 1px dotted rgba(256, 256, 256, 0.7);
   padding: 0.5rem 0;
 }
 
 .Track-action {
   cursor: pointer;
   padding: 0.5rem;
-  font-size: 1.05rem;
+  font-size: 1rem;
   transition: color 0.25s;
   border: 0;
   background-color: rgb(0, 0, 0);
@@ -32,7 +32,7 @@
 }
 
 .Track-information p {
-  font-size: 0.83rem;
+  font-size: 0.8rem;
   font-weight: 300;
 }
