+++ src/components/Track/Track.js
@@ -19,16 +19,24 @@ class Track extends Component {
 
   renderAction() {
     if (this.props.isRemoval) {
-      return <button className="Track-action" onClick={this.removeTrack}>-</button>;
+      return (
+        <button className="Track-action" onClick={this.removeTrack}>
+          -
+        </button>
+      );
     }
-    return <button className="Track-action" onClick={this.addTrack}>+</button>;
+    return (
+      <button className="Track-action" onClick={this.addTrack}>
+        +
+      </button>
+    );
   }
 
   render() {
     return (
       <div className="Track">
         <div className="Track-information">
-          <h3>{this.props.track.name}</h3>
+          <h3 style={{ fontSize: '1rem' }}>{this.props.track.name}</h3>
           <p>
             {this.props.track.artist} | {this.props.track.album}
           </p>
